requirements.txt
googlesearch-python==1.3.0
tqdm==4.67.1